Moved the unconditional ok return in the compiler reaxt_webpack to the task webpack.compile
A task return value is unused by mix, but this is where we can determine
if it ran successfully or not. Moving the final value from the main
compiler to the task manually called minimize the amount of change, and
the amount of  redundant case analysis to be done inside the compiler.